requirements.txt
TikTokApi>=7.0.0
python-dotenv>=1.0.0
certifi>=2024.0.0
flask>=3.0.0
gunicorn>=21.0.0
